+++ template_variables/src/epilot_template_variables/sdkconfiguration.py
@@ -26,10 +26,10 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "1.0.0"
-    sdk_version: str = "1.12.0"
-    gen_version: str = "2.466.0"
+    sdk_version: str = "1.12.1"
+    gen_version: str = "2.467.4"
     user_agent: str = (
-        "speakeasy-sdk/python 1.12.0 2.466.0 1.0.0 epilot-template-variables"
+        "speakeasy-sdk/python 1.12.1 2.467.4 1.0.0 epilot-template-variables"
     )
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None