+++ template_variables/RELEASES.md
@@ -211,3 +211,11 @@ Based on:
 - Speakeasy CLI 1.448.0 (2.466.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.12.0] template_variables
+
+## 2024-11-28 00:30:33
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.448.2 (2.467.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.12.1] template_variables